project04_assignments/homework_projects/02_lists/03_erase_canvas/pyproject.toml
[project]
name = "03-erase-canvas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "graphics-py>=5.0.1.post1",
]
